requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arelle==2.2
arelle-release==2.37.6
attrs==25.3.0
beautifulsoup4==4.13.4
bottle==0.13.3
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.2.0
coverage==7.8.0
distro==1.9.0
edgar==5.6.3
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-secret-manager==2.23.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
holidays==0.72
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
numpy==2.2.5
oauthlib==3.2.2
openai==1.77.0
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
pyrate-limiter==3.7.0
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-cov==6.1.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
sec-edgar-api==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
trio==0.30.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zstandard==0.23.0
